+++ app/code/local/Autocompleteplus/Autosuggest/Helper/Catalogsearch.php
@@ -18,7 +18,26 @@
  * @license   Open Software License (OSL 3.0)*
  * @link      http://opensource.org/licenses/osl-3.0.php
  */
- 
+
+/**
+ * Autocompleteplus_Autosuggest_Helper_Catalogsearch
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @category Mage
+ *
+ * @package   Instantsearchplus
+ * @author    Fast Simon <info@instantsearchplus.com>
+ * @copyright 2017 Fast Simon (http://www.instantsearchplus.com)
+ * @license   Open Software License (OSL 3.0)*
+ * @link      http://opensource.org/licenses/osl-3.0.php
+ */
  class Autocompleteplus_Autosuggest_Helper_Catalogsearch extends Mage_CatalogSearch_Helper_Data
  {
      public function getResultUrl($query = null)